refactoring of old_forces

+++ src/cpu-new/include/common.hpp
@@ -3,13 +3,15 @@
 #include <iomanip>
 #include <omp.h>
 
+#include <iostream>
+
 const int J=10;
 const int INIT_PARTICLE=0;
 const float RADIUS_MASS_PORCENTAGE=0.2;
 const double E=1e-4;
 const double E2=1e-8;
-const float ETA_S=0.01;
-const float ETA_N=0.01;
+const double  ETA_S=0.01;
+const double  ETA_N=0.01;
 const double D_TIME_MIN=1.1920928955078125e-07;
 const double D_TIME_MAX=0.125;
 
@@ -67,5 +69,7 @@ typedef struct file_data
     double r[3];
     double v[3];
 } file_data;
+
+
 
 #endif
