+++ src/cpu-new/include/extra_utils.hpp
@@ -1,10 +1,11 @@
 #ifndef EXTRAUTILS_HPP
 #define EXTRAUTILS_HPP
-#include <cmath>
 #include "common.hpp"
+#include <cmath>
 
-double  get_magnitude(double x, double y, double z);
-double get_timestep_normal(int i, double4 *a2, double4 *a3, double *dt, Forces *f, float eta);
+double get_magnitude(double x, double y, double z);
+double get_timestep_normal(int i, double4 *a2, double4 *a3, double *dt, Forces *f,
+                           double eta);
 double normalize_dt(double new_dt, double old_dt, double t, int i);
 
 #endif
